[Android] api UI 연동 #329 (#349)
* feat: Home mvi 구현

* feat: Home Route에 인자 추가

* feat: result로 리턴 타입 수정

* feat: 오타 수정

* feat: 기본 인자 추가

* feat: 방 타입 이넘 클래스 생성

* feat: Immutable 래퍼 클래스 생성

* feat: 백스택 타입 수정

* feat: Home에 로직 적용

* feat: Home 네비게이션 인자 추가

* feat: state Update 하는 로직 추가

* style: lint 적용

* feat: 타입 수정

* feat: 경로 오타 추가

* feat: getter함수 제거, 접근제한자 추가

* feat: 오타 수정된 것으로 적용

+++ src/android/bbebig-discod-clone/presentation/src/main/java/com/smilegate/bbebig/presentation/ui/home/model/Server.kt
@@ -3,13 +3,13 @@ package com.smilegate.bbebig.presentation.ui.home.model
 import com.smilegate.bbebig.domain.model.ServerDomainModel
 
 data class Server(
-    val serverId: Int,
+    val serverId: Long,
     val serverName: String,
     val serverImageUrl: String?,
 )
 
 fun ServerDomainModel.toUiModel(): Server = Server(
-    serverId = serverId,
+    serverId = serverId.toLong(),
     serverName = serverName,
     serverImageUrl = serverImageUrl,
 )
